pyproject.toml
[tool.poetry]
name = "whiteboard-backend-2"
version = "0.1.0"
description = ""
authors = ["Nirlep5252 <nirlep5252@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
fastapi = "^0.104.1"
python-keycloak = "^3.3.0"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
python-dotenv = "^1.0.0"
databases = "^0.8.0"
icecream = "^2.1.3"
websockets = "^12.0"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
dataclasses-json = "^0.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
